// frontend/packages/renderer/src/generated/remotion-shapes-props.ts
export const REMOTION_SHAPES_SPECIFIC_PROP_MAP: Record<string, readonly string[]> = {
  "Circle": [
    "radius"
  ],
  "Ellipse": [
    "rx",
    "ry"
  ],
  "Heart": [
    "aspectRatio",
    "bottomRoundnessAdjustment",
    "depthAdjustment",
    "height"
  ],
  "Pie": [
    "closePath",
    "counterClockwise",
    "progress",
    "radius",
    "rotation"
  ],
  "Polygon": [
    "cornerRadius",
    "edgeRoundness",
    "points",
    "radius"
  ],
  "Rect": [
    "cornerRadius",
    "edgeRoundness",
    "height",
    "width"
  ],
  "Star": [
    "cornerRadius",
    "edgeRoundness",
    "innerRadius",
    "outerRadius",
    "points"
  ],
  "Triangle": [
    "cornerRadius",
    "direction",
    "edgeRoundness",
    "length"
  ]
} as const;